refactor: split sva-mainserver service internals
